I added all of my edited files into a new appfile so there would be less conflict
