+++ .idea/workspace.xml
@@ -1,11 +1,18 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project version="4">
   <component name="ChangeListManager">
-    <list default="true" id="076d4416-8162-4cce-9cbe-e1e370322afa" name="Default Changelist" comment="fixed up some functions and view and update html">
-      <change afterPath="$PROJECT_DIR$/src/LOR/templates/LOR/uploadview.html" afterDir="false" />
+    <list default="true" id="076d4416-8162-4cce-9cbe-e1e370322afa" name="Default Changelist" comment="I added a view to be able to view all of the uploads, and also be able to delete them">
+      <change afterPath="$PROJECT_DIR$/src/Upload/form.py" afterDir="false" />
+      <change afterPath="$PROJECT_DIR$/src/Upload/urls.py" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/.idea/workspace.xml" beforeDir="false" afterPath="$PROJECT_DIR$/.idea/workspace.xml" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/LOR/form.py" beforeDir="false" afterPath="$PROJECT_DIR$/src/LOR/form.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/LOR/models.py" beforeDir="false" afterPath="$PROJECT_DIR$/src/LOR/models.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/LOR/templates/LOR/upload.html" beforeDir="false" afterPath="$PROJECT_DIR$/src/Upload/templates/Upload/upload.html" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/LOR/templates/LOR/uploadview.html" beforeDir="false" afterPath="$PROJECT_DIR$/src/Upload/templates/Upload/uploadview.html" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/src/LOR/urls.py" beforeDir="false" afterPath="$PROJECT_DIR$/src/LOR/urls.py" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/src/LOR/views.py" beforeDir="false" afterPath="$PROJECT_DIR$/src/LOR/views.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/ReccomendationPortal/settings.py" beforeDir="false" afterPath="$PROJECT_DIR$/src/ReccomendationPortal/settings.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/ReccomendationPortal/urls.py" beforeDir="false" afterPath="$PROJECT_DIR$/src/ReccomendationPortal/urls.py" afterDir="false" />
     </list>
     <option name="SHOW_DIALOG" value="false" />
     <option name="HIGHLIGHT_CONFLICTS" value="true" />
@@ -39,6 +46,11 @@
     <property name="SHARE_PROJECT_CONFIGURATION_FILES" value="true" />
     <property name="WebServerToolWindowFactoryState" value="false" />
   </component>
+  <component name="RecentsManager">
+    <key name="MoveFile.RECENT_KEYS">
+      <recent name="C:\Users\thuan\PycharmProjects\Reccomendation-Portal2\src\Upload\templates\Upload" />
+    </key>
+  </component>
   <component name="SpellCheckerSettings" RuntimeDictionaries="0" Folders="0" CustomDictionaries="0" DefaultDictionary="application-level" UseSingleDictionary="true" transferred="true" />
   <component name="TaskManager">
     <task active="true" id="Default" summary="Default task">
@@ -71,7 +83,14 @@
       <option name="project" value="LOCAL" />
       <updated>1618269144128</updated>
     </task>
-    <option name="localTasksCounter" value="4" />
+    <task id="LOCAL-00004" summary="I added a view to be able to view all of the uploads, and also be able to delete them">
+      <created>1618342113896</created>
+      <option name="number" value="00004" />
+      <option name="presentableId" value="LOCAL-00004" />
+      <option name="project" value="LOCAL" />
+      <updated>1618342113896</updated>
+    </task>
+    <option name="localTasksCounter" value="5" />
     <servers />
   </component>
   <component name="TypeScriptGeneratedFilesManager">
@@ -92,6 +111,7 @@
   <component name="VcsManagerConfiguration">
     <MESSAGE value="added some function to submit loaded form to a static file on local computer called media" />
     <MESSAGE value="fixed up some functions and view and update html" />
-    <option name="LAST_COMMIT_MESSAGE" value="fixed up some functions and view and update html" />
+    <MESSAGE value="I added a view to be able to view all of the uploads, and also be able to delete them" />
+    <option name="LAST_COMMIT_MESSAGE" value="I added a view to be able to view all of the uploads, and also be able to delete them" />
   </component>
 </project>
