+++ src/Upload/form.py
@@ -1,0 +1,7 @@
+from django import forms
+from .models import Upload
+
+class Uploadform(forms.ModelForm):
+    class Meta:
+        model = Upload
+        fields = ('name', 'type', 'pdf')
